jarvis_webapp/requirements.txt
flask==3.0.3
flask-socketio==5.3.6
flask-cors==4.0.0
eventlet==0.35.2
psutil==5.9.8
requests==2.31.0
wikipedia-api==0.6.0
textblob==0.18.0
pyttsx3==2.90
SpeechRecognition==3.10.4
pyaudio==0.2.14
anthropic==0.28.0
python-dotenv==1.0.1
colorama==0.4.6
pyautogui==0.9.54
